fase-07b/migrations: repair history + aplicar 6 migrations + regen types
Repair migration_history (metadata-only, no schema change):
- reverted: 071815-074224 (7 entries remote-only renombradas en sesion previa)
- applied: 080000-086000 (7 entries local que ya estaban en remote schema)

db push aplica las 6 migrations realmente nuevas:
- 087000_str_sentiment_anthropic_budget (api_budgets.anthropic + aggregate_zone_sentiment + monthly_anthropic_spend)
- 088000_str_invisible_hotels (table + detect_invisible_hotel_candidates fn)
- 089000_str_zone_aqi_aggregate (zone_aqi_summary fn)
- 090000_str_host_migrations (table + market_migration_alert_pct fn)
- 091000_str_pricing_calendar (str_events_calendar + str_pricing_overrides + seed eventos MX 2026)
- 092000_str_reports (table + RLS owner-or-admin)

Fixes SQL detectados durante apply:
- 087000 aggregate_zone_sentiment: extract(day from date-date) invalido en PG;
  reemplazado por (current_date - posted_at)::numeric.
- 088000 detect_invisible_hotel_candidates: window function dentro de aggregate
  no permitido; refactor a CTE centroids + CTE with_radius separadas.

npm run db:types regenera shared/types/database.ts desde remote actualizado.
Tipos manuales previos (ampliaciones in-line de database.ts) ahora redundantes
pero compatibles. Suite: 647 passing | 2 skipped, tsc clean, lint clean.

+++ supabase/migrations/20260419087000_str_sentiment_anthropic_budget.sql
@@ -83,8 +83,8 @@ as $$
       topics,
       exp(
         -ln(2)
-        * extract(day from (current_date - posted_at))
-        / nullif(p_decay_half_life_days, 0)
+        * (current_date - posted_at)::numeric
+        / nullif(p_decay_half_life_days, 0)::numeric
       ) as weight
     from reviews_in_scope
   ),
